EfficientNetV2 Models

Constructs EfficientNetV2 model architectures as described in EfficientNetV2: Smaller Models and Faster Training.

model_efficientnet_v2_s(pretrained = FALSE, progress = TRUE, ...)

model_efficientnet_v2_m(pretrained = FALSE, progress = TRUE, ...)

model_efficientnet_v2_l(pretrained = FALSE, progress = TRUE, ...)

Arguments

pretrained

(bool): If TRUE, returns a model pre-trained on ImageNet.

progress

(bool): If TRUE, displays a progress bar of the download to stderr.

...

Other parameters passed to the model implementation, such as num_classes to change the output dimension.

Functions

model_efficientnet_v2_s(): EfficientNetV2-S model

model_efficientnet_v2_m(): EfficientNetV2-M model

model_efficientnet_v2_l(): EfficientNetV2-L model

Task

Image classification with 1000 output classes by default (ImageNet).

Input Format

The models expect input tensors of shape (batch_size, 3, H, W). Typical values for H and W are 384 for V2-S, 480 for V2-M, and 512 for V2-L.

Variants

Model	Resolution	Params (M)	GFLOPs	Top-1 Acc.

V2-S	384	24	8.4	83.9

V2-M	480	55	24	85.1

V2-L	512	119	55	85.7

See also

Examples