Crops the given image at the center

The image can be a Magick Image or a torch Tensor, in which case it is expected to have [..., H, W] shape, where ... means an arbitrary number of leading dimensions.

transform_center_crop(img, size)

Arguments

img
A magick-image, array or torch_tensor.

size
(sequence or int): Desired output size of the crop. If size is an int instead of sequence like (h, w), a square crop (size, size) is made. If provided a tuple or list of length 1, it will be interpreted as (size, size).

See also