MNIST dataset

Prepares the MNIST dataset and optionally downloads it.

mnist_dataset(
  root,
  train = TRUE,
  transform = NULL,
  target_transform = NULL,
  download = FALSE
)

Arguments

root
(string): Root directory of dataset where MNIST/processed/training.pt and MNIST/processed/test.pt exist.

train
(bool, optional): If True, creates dataset from training.pt, otherwise from test.pt.

transform
(callable, optional): A function/transform that takes in an PIL image and returns a transformed version. E.g, transforms.RandomCrop.

target_transform
(callable, optional): A function/transform that takes in the target and transforms it.

download
(bool, optional): If true, downloads the dataset from the internet and puts it in root directory. If dataset is already downloaded, it is not downloaded again.